qcam: Provide initial icon buttons "Play/Stop"
Provide Quit, Play, Pause, Stop icons.

Utilise the provided QT resources to present icons for the toolbar.

Update the Quit button with a 'cross', and implement Play/Pause/Stop
buttons.

'Pause' is a no-op currently and likely could be removed, but I wanted
an ability to distinguish between holding the stream and restarting it.

Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>